Booking reminder/requirements.txt
functions-framework==3.*
google-cloud-firestore
firebase_admin
google-cloud-pubsub